backend/requirements/base.txt
Django>=5.0,<5.1
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
djangorestframework-simplejwt>=5.3,<6.0
drf-spectacular>=0.27,<1.0
django-filter>=23.5,<25.0
channels>=4.0,<5.0
channels-redis>=4.1,<5.0
daphne>=4.0,<5.0
celery>=5.3,<6.0
redis>=5.0,<6.0
WeasyPrint>=61.0,<63.0
openpyxl>=3.1,<4.0
Pillow>=10.0,<11.0
psycopg2-binary>=2.9,<3.0
python-dotenv>=1.0,<2.0
